Create simple routing middleware

+++ routes.php
@@ -5,20 +5,20 @@ $router->get('/', 'HomeController@index');
 
 // Listing Display Routes
 $router->get('/listings', 'ListingsController@index');
-$router->get('/listings/create', 'ListingsController@create');
+$router->get('/listings/create', 'ListingsController@create', ['auth']);
 $router->get('/listings/{id}', 'ListingsController@show');
-$router->get('/listings/{id}/edit', 'ListingsController@edit');
+$router->get('/listings/{id}/edit', 'ListingsController@edit', ['auth']);
 
 // Listing CRUD Routes
-$router->post('/listings', 'ListingsController@store');
-$router->put('/listings/{id}', 'ListingsController@update');
-$router->delete('/listings/{id}', 'ListingsController@destroy');
+$router->post('/listings', 'ListingsController@store', ['auth']);
+$router->put('/listings/{id}', 'ListingsController@update', ['auth']);
+$router->delete('/listings/{id}', 'ListingsController@destroy', ['auth']);
 
 // Authentication Display Routes
-$router->get('/register', 'UserController@register');
-$router->get('/login', 'UserController@login');
+$router->get('/register', 'UserController@register', ['guest']);
+$router->get('/login', 'UserController@login', ['guest']);
 
 // Authentication CRUD Routes
-$router->post('/register', 'UserController@store');
-$router->post('/login', 'UserController@authenticate');
-$router->post('/logout', 'UserController@logout');
+$router->post('/register', 'UserController@store', ['guest']);
+$router->post('/login', 'UserController@authenticate', ['guest']);
+$router->post('/logout', 'UserController@logout', ['auth']);
